refactor: replace knowledge_graph vector infrastructure with shared module imports
- Replace vector_index.py (~250 lines) with 5-line re-export shim from akgentic.tool.vector
- Update kg_actor.py to import EmbeddingService/VectorEntry/VectorIndex from akgentic.tool.vector
- Update __init__.py to re-export from akgentic.tool.vector and alias _check_kg_dependencies
- Update test mock targets from module-level openai to lazy import patching

Relates to #21
